% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_precrec.R
\name{as_precrec}
\alias{as_precrec}
\alias{as_precrec.PredictionClassif}
\alias{as_precrec.ResampleResult}
\alias{as_precrec.BenchmarkResult}
\title{Convert to 'precrec' Format}
\usage{
as_precrec(object)

\method{as_precrec}{PredictionClassif}(object)

\method{as_precrec}{ResampleResult}(object)

\method{as_precrec}{BenchmarkResult}(object)
}
\arguments{
\item{object}{:: \code{any}\cr
Object to convert.}
}
\value{
Object as created by \code{\link[precrec:mmdata]{precrec::mmdata()}}.
}
\description{
Converts to a format which is understood by \code{\link[precrec:evalmod]{precrec::evalmod()}} of package \CRANpkg{precrec}.
}
