% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{autoplot.Filter}
\alias{autoplot.Filter}
\title{Plot for Filter Scores}
\usage{
\method{autoplot}{Filter}(object, n = Inf, ...)
}
\arguments{
\item{object}{(\link[mlr3:Filter]{mlr3::Filter}).}

\item{n}{(\code{integer(1)})\cr
Only include the first \code{n} features with highest importance.
Defaults to all features.}

\item{...}{(\code{any}):
Additional argument, possibly passed down to the underlying plot functions.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3filters:Filter]{mlr3filters::Filter}.
}
\examples{
library(mlr3)
library(mlr3filters)
task = mlr_tasks$get("mtcars")
f = mlr_filters$get("correlation")
f$calculate(task)

head(fortify(f))
autoplot(f, n = 5)
}
