% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimInstanceSingleCrit.R
\name{autoplot.OptimInstanceSingleCrit}
\alias{autoplot.OptimInstanceSingleCrit}
\title{Plot for OptimInstanceSingleCrit}
\usage{
\method{autoplot}{OptimInstanceSingleCrit}(
  object,
  type = "marginal",
  cols_x = NULL,
  trafo = FALSE,
  learner = mlr3::lrn("regr.ranger"),
  grid_resolution = 100,
  ...
)
}
\arguments{
\item{object}{(\link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit}.}

\item{type}{(\code{character(1)}):
Type of the plot. Available choices:
\itemize{
\item \code{"marginal"}: scatter plots of x versus y The
colour of the points shows the batch number.
\item \code{"performance"}: scatter plots of batch number versus y
\item \code{"parameter"}: scatter plots of batch number versus input. The
colour of the points shows the y values.
\item \code{"parallel"} parallel coordinates plot. x values are rescaled by
\code{(x - mean(x)) / sd(x)}.
\item \code{"points"} - scatter plot of two x dimensions versus y The
colour of the points shows the y values.
\item \code{"surface"}: surface plot of two x dimensions versus y values.
The y values are interpolated with the supplied
\link[mlr3:Learner]{mlr3::Learner}.
\item \code{"pairs"}: plots all x and y values against each other.
}}

\item{cols_x}{(\code{character()})\cr
Column names of x values. By default, all untransformed
x values from the search space are plotted. Transformed hyperparameters are prefixed with
\code{x_domain_}.}

\item{trafo}{(\code{logical(1)})\cr
Determines if untransformed (\code{FALSE}) or transformed (\code{TRUE})
x values are plotted.}

\item{learner}{(\link[mlr3:Learner]{mlr3::Learner})\cr
Regression learner used to interpolate the data of the surface plot.}

\item{grid_resolution}{(\code{numeric()})\cr
Resolution of the surface plot.}

\item{...}{(\code{any}):
Additional arguments, possibly passed down to the underlying plot functions.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit}.
}
\examples{
if (requireNamespace("bbotk") && requireNamespace("patchwork")) {
  library(bbotk)
}
}
