% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionClust.R
\name{autoplot.PredictionClust}
\alias{autoplot.PredictionClust}
\title{Plot for PredictionClust}
\usage{
\method{autoplot}{PredictionClust}(object, task, type = "scatter", ...)
}
\arguments{
\item{object}{(\link[mlr3:PredictionClust]{mlr3::PredictionClust}).}

\item{task}{(\link[mlr3:TaskClust]{mlr3::TaskClust}).}

\item{type}{(character(1)):\cr
Type of the plot. See description.}

\item{...}{(\code{any}):
Additional arguments, passed down to the respective \code{geom}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:PredictionClust]{mlr3::PredictionClust}, depending on argument \code{type}:
\itemize{
\item \code{"scatter"} (default): scatterplot with correlation values
and colored cluster assignments.
}
}
\examples{
library(mlr3)
library(mlr3cluster)
library(mlr3viz)

task = tsk("usarrests")
learner = lrn("clust.kmeans", centers = 3)
object = learner$train(task)$predict(task)

head(fortify(object))
autoplot(object, task)
}
