% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{autoplot.ResampleResult}
\alias{autoplot.ResampleResult}
\title{Plot for ResampleResult}
\usage{
\method{autoplot}{ResampleResult}(object, type = "boxplot", measure = NULL, ...)
}
\arguments{
\item{object}{(\link[mlr3:ResampleResult]{mlr3::ResampleResult}).}

\item{type}{(character(1)):\cr
Type of the plot.}

\item{measure}{(\link[mlr3:Measure]{mlr3::Measure}).}

\item{...}{(\code{any}):
Additional arguments, passed down to the respective \code{geom}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:ResampleResult]{mlr3::ResampleResult}, depending on argument \code{type}:
\itemize{
\item \code{"boxplot"} (default): Boxplot of performance measures.
\item \code{"histogram"}: Histogram of performance measures.
}
}
\examples{
library(mlr3)
task = mlr_tasks$get("iris")
learner = mlr_learners$get("classif.rpart")
resampling = mlr_resamplings$get("cv")
object = resample(task, learner, resampling)

head(fortify(object))
autoplot(object)
autoplot(object, type = "histogram", binwidth = 0.01)
}
