% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv.R
\name{autoplot.TaskSurv}
\alias{autoplot.TaskSurv}
\title{Plot for Survival Tasks}
\usage{
\method{autoplot}{TaskSurv}(object, type = "target", ...)
}
\arguments{
\item{object}{(\link[mlr3survival:TaskSurv]{mlr3survival::TaskSurv}).}

\item{type}{(\code{character(1)}):
Type of the plot. Available choices:
\itemize{
\item \code{"target"}: bar plot of target variable (default).
\item \code{"duo"}: Passes data and additional arguments down to \link[GGally:ggduo]{GGally::ggduo}.
\code{columnsX} is target, \code{columnsY} is features.
\item \code{"pairs"}: Passes data and additional arguments down to \link[GGally:ggpairs]{GGally::ggpairs}.
Color is set to target column.
}}

\item{...}{(any):
Additional argument, possibly passed down to the underlying plot functions.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3survival:TaskSurv]{mlr3survival::TaskSurv}.
}
\examples{
library(mlr3)
library(mlr3survival)
task = mlr_tasks$get("lung")

head(fortify(task))
autoplot(task)
autoplot(task, strata = "sex")
autoplot(task, type = "duo")
}
