% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{fortify}
\alias{fortify}
\title{Re-export of \code{fortify}

See \code{\link[ggplot2:fortify]{ggplot2::fortify()}}.}
\description{
Re-export of \code{fortify}

See \code{\link[ggplot2:fortify]{ggplot2::fortify()}}.
}
\keyword{internal}
