% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3viz-package}
\alias{mlr3viz}
\alias{mlr3viz-package}
\title{mlr3viz: Vizualizations for mlr3}
\description{
Provides `autoplot()` like visualizations for 'mlr3' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3viz.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3viz}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3viz/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

}
