% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{predict_grid}
\alias{predict_grid}
\title{Generates a data.table of evenly distributed points.}
\usage{
predict_grid(learners, task, grid_points, expand_range)
}
\arguments{
\item{learners}{list of trained learners, each learner belongs to one
resampling iteration}

\item{task}{the task all learners are trained on}

\item{grid_points}{(int): see sequenize}

\item{expand_range}{see sequenize}
}
\description{
For each point we have the predicted class / regression value  in column
response. If the learner predicts probabilities, a column ".prob.response" is
added that contains the probability of the predicted class
}
