% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver Operatoring Characteristic Plots}
\usage{
roc(prediction)
}
\arguments{
\item{prediction}{:: `\link[mlr3:PredictionClassif]{mlr3::PredictionClassif}.}
}
\value{
Object of type \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
\strong{This function is not finished yet. Do not use.}

Takes a \link[mlr3:PredictionClassif]{mlr3::PredictionClassif} for a binary classification problem and
plots a ROC curve.
Internally calls the function \code{\link[precrec:evalmod]{precrec::evalmod()}} from package \CRANpkg{precrec}.
}
\examples{
library(mlr3)
task = mlr_tasks$get("spam")
lrn = mlr_learners$get("classif.rpart")
lrn$predict_type = "prob"
prediction = lrn$train(task)$predict(task)
p = roc(prediction)
}
\keyword{internal}
