% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\name{autoplot.BenchmarkResult}
\alias{autoplot.BenchmarkResult}
\title{Plot for BenchmarkResult}
\usage{
\method{autoplot}{BenchmarkResult}(object, type = "boxplot", measure = NULL, ...)
}
\arguments{
\item{object}{(\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}).}

\item{type}{(character(1)):\cr
Type of the plot.}

\item{measure}{(\link[mlr3:Measure]{mlr3::Measure}).}

\item{...}{(\code{any}):
Currently ignored.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}, depending on argument \code{type}:
\itemize{
\item \code{"boxplot"} (default): Boxplots of performance measures, one box per \link[mlr3:Learner]{mlr3::Learner} and one facet per \link[mlr3:Task]{mlr3::Task}.
}
}
\examples{
library(mlr3)
tasks = mlr_tasks$mget(c("iris", "pima", "sonar"))
learner = mlr_learners$mget(c("classif.featureless", "classif.rpart"))
resampling = mlr_resamplings$mget("cv")
object = benchmark(benchmark_grid(tasks, learner, resampling))

head(fortify(object))
autoplot(object)
}
