% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\name{autoplot.BenchmarkResult}
\alias{autoplot.BenchmarkResult}
\title{Plot for BenchmarkResult}
\usage{
\method{autoplot}{BenchmarkResult}(object, type = "boxplot",
  by = "task_id", ...)
}
\arguments{
\item{object}{[\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}].}

\item{type}{[character(1)]: Type of the plot.}

\item{by}{[character(1)]: Group plots by this variable.}

\item{...}{[any]: Additional arguments, passed down to the respective \code{geom}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}.
}
\examples{
library(mlr3)
tasks = mlr_tasks$mget(c("iris", "pima", "sonar"))
learner = mlr_learners$mget(c("classif.featureless", "classif.rpart"))
resampling = mlr_resamplings$mget("cv")
object = benchmark(tasks, learner, resampling)

library(ggplot2)
fortify(object)
autoplot(object)
}
