% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\name{autoplot.BenchmarkResult}
\alias{autoplot.BenchmarkResult}
\title{Plot for BenchmarkResult}
\usage{
\method{autoplot}{BenchmarkResult}(object, measure = NULL)
}
\arguments{
\item{object}{[\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}].}

\item{measure}{[\link[mlr3:Measure]{mlr3::Measure}]. If NULL, defaults to first measure, i.e.
\code{object$measures$measure[[1]]}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}.
}
\examples{
library(mlr3)
tasks = mlr_tasks$mget(c("iris", "pima", "sonar"))
learner = mlr_learners$mget(c("classif.featureless", "classif.rpart"))
resampling = mlr_resamplings$mget("cv")
object = benchmark(tasks, learner, resampling)

library(ggplot2)
fortify(object)
autoplot(object)
}
