% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifCVGlmnet.R,
%   R/LearnerClassifGlmnet.R, R/LearnerRegrCVGlmnet.R, R/LearnerRegrGlmnet.R
\name{autoplot.LearnerClassifCVGlmnet}
\alias{autoplot.LearnerClassifCVGlmnet}
\alias{autoplot.LearnerClassifGlmnet}
\alias{autoplot.LearnerRegrCVGlmnet}
\alias{autoplot.LearnerRegrGlmnet}
\title{Plot for LearnerClassifGlmnet / LearnerRegrGlmnet / LearnerClassifCVGlmnet / LearnerRegrCVGlmnet}
\usage{
\method{autoplot}{LearnerClassifCVGlmnet}(object, ...)

\method{autoplot}{LearnerClassifGlmnet}(object, ...)

\method{autoplot}{LearnerRegrCVGlmnet}(object, ...)

\method{autoplot}{LearnerRegrGlmnet}(object, ...)
}
\arguments{
\item{object}{(\link[mlr3learners:mlr_learners_classif.glmnet]{mlr3learners::LearnerClassifGlmnet} | \link[mlr3learners:mlr_learners_regr.glmnet]{mlr3learners::LearnerRegrGlmnet} |
\link[mlr3learners:mlr_learners_regr.cv_glmnet]{mlr3learners::LearnerRegrCVGlmnet} | \link[mlr3learners:mlr_learners_regr.cv_glmnet]{mlr3learners::LearnerRegrCVGlmnet}).}

\item{...}{(\code{any}):
Additional arguments, passed down to \code{\link[ggparty:autoplot.party]{ggparty::autoplot.party()}}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Visualizations for \link[mlr3learners:mlr_learners_classif.glmnet]{mlr3learners::mlr_learners_classif.glmnet}, \link[mlr3learners:mlr_learners_regr.glmnet]{mlr3learners::mlr_learners_regr.glmnet},
\link[mlr3learners:mlr_learners_classif.cv_glmnet]{mlr3learners::mlr_learners_classif.cv_glmnet} and \link[mlr3learners:mlr_learners_regr.cv_glmnet]{mlr3learners::mlr_learners_regr.cv_glmnet}
using the package \CRANpkg{ggfortify}.

Note that learner-specific plots are experimental and subject to change.
}
\examples{
\dontrun{
library(mlr3)
library(mlr3viz)
library(mlr3learners)

# classification
task = tsk("sonar")
learner = lrn("classif.glmnet")
learner$train(task)
autoplot(learner)

# regression
task = tsk("mtcars")
learner = lrn("regr.glmnet")
learner$train(task)
autoplot(learner)
}
}
\references{
Tang Y, Horikoshi M, Li W (2016).
\dQuote{ggfortify: Unified Interface to Visualize Statistical Result of Popular R Packages.}
\emph{The R Journal}, \bold{8}(2), 474--485.
\doi{10.32614/RJ-2016-060}.
}
