% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif.R
\name{autoplot.TaskClassif}
\alias{autoplot.TaskClassif}
\title{Plot for Classification Tasks}
\usage{
\method{autoplot}{TaskClassif}(object, type = "target")
}
\arguments{
\item{object}{(\link[mlr3:TaskClassif]{mlr3::TaskClassif}).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:TaskClassif]{mlr3::TaskClassif}.
}
\examples{
library(mlr3)
object = mlr_tasks$get("pima")

library(ggplot2)
autoplot(object)
autoplot(object, type = "pairs")
autoplot(object, type = "duo")
}
