% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{autoplot.TaskRegr}
\alias{autoplot.TaskRegr}
\title{Plot for Regression Tasks}
\usage{
\method{autoplot}{TaskRegr}(object, type = "target")
}
\arguments{
\item{object}{(\link[mlr3:TaskRegr]{mlr3::TaskRegr}).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:TaskRegr]{mlr3::TaskRegr}.
}
\examples{
library(mlr3)
object = mlr_tasks$get("mtcars")

library(ggplot2)
autoplot(object)
autoplot(object, type = "pairs")
}
