% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{autoplot.TaskRegr}
\alias{autoplot.TaskRegr}
\title{Plot for Regression Tasks}
\usage{
\method{autoplot}{TaskRegr}(object, type = "target", ...)
}
\arguments{
\item{object}{(\link[mlr3:TaskRegr]{mlr3::TaskRegr}).}

\item{type}{(\code{character(1)}):
Type of the plot. Available choices:
\itemize{
\item \code{"target"}: bar plot of target variable (default).
\item \code{"pairs"}: Passes data and additional arguments down to \link[GGally:ggpairs]{GGally::ggpairs}.
Color is set to target column.
}}

\item{...}{(any):
Additional argument, possibly passed down to the underlying plot functions.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3:TaskRegr]{mlr3::TaskRegr}.
}
\examples{
library(mlr3)
task = mlr_tasks$get("mtcars")
task$select(c("mpg", "am", "carb"))

head(fortify(task))
autoplot(task)
autoplot(task, type = "pairs")
}
