% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3viz-package}
\alias{mlr3viz}
\alias{mlr3viz-package}
\title{mlr3viz: Vizualizations for mlr3}
\description{
Provides visualizations for mlr3 objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlrviz.mlr-org.com}
  \item Report bugs at \url{https://github.com/mlr-org/mlrviz/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

}
